refactor by moving functions
